clean up before submitting PR
